refactor: separate registration(sign-up) from authentication
