+++ lib/main.dart
@@ -6,9 +6,9 @@ import 'package:path_provider/path_provider.dart';
 import 'package:simple_notes_app/router/app_router.dart';
 import 'package:simple_notes_app/service/service.dart';
 import 'package:simple_notes_app/theme/app_theme.dart';
-import 'package:simple_notes_app/view/authentication/sign_in/bloc/sign_in_bloc.dart';
-import 'package:simple_notes_app/view/authentication/sign_up/bloc/sign_up_bloc.dart';
+import 'package:simple_notes_app/view/authentication/authentication.dart';
 import 'package:simple_notes_app/view/profile/bloc/auth_bloc.dart';
+import 'package:simple_notes_app/view/registration/registration.dart';
 import 'package:simple_notes_app/view/settings/theme_bloc/theme_bloc.dart';
 
 Future<void> main() async {
