+++ lib/service/service.dart
@@ -1,4 +1,4 @@
 export 'package:simple_notes_app/service/src/authentication/sign_in_remote_service.dart';
-export 'package:simple_notes_app/service/src/authentication/sign_up_remote_service.dart';
 export 'package:simple_notes_app/service/src/notes_service.dart';
+export 'package:simple_notes_app/service/src/registration/sign_up_remote_service.dart';
 export 'package:simple_notes_app/service/src/user_account/user_account_remote_service.dart';
