+++ lib/view/authentication/authentication.dart
@@ -1,2 +1,2 @@
+export 'package:simple_notes_app/view/authentication/sign_in/bloc/sign_in_bloc.dart';
 export 'package:simple_notes_app/view/authentication/sign_in/sign_in_screen.dart';
-export 'package:simple_notes_app/view/authentication/sign_up/sign_up_screen.dart';
