+++ lib/view/registration/registration.dart
@@ -1,0 +1,2 @@
+export 'sign_up/bloc/sign_up_bloc.dart';
+export 'sign_up/sign_up_screen.dart';
