+++ lib/view/view.dart
@@ -5,5 +5,6 @@ export 'package:simple_notes_app/view/notes/edit_note_screen.dart';
 export 'package:simple_notes_app/view/notes/note_details_screen.dart';
 export 'package:simple_notes_app/view/notes/notes_screen.dart';
 export 'package:simple_notes_app/view/profile/profile_screen.dart';
+export 'package:simple_notes_app/view/registration/registration.dart';
 export 'package:simple_notes_app/view/settings/settings_screen.dart';
 export 'package:simple_notes_app/view/src/bottom_navigation_scaffold.dart';
